feat(Home.tsx): add Home component for audio transcription functionality refactor(page.tsx): replace inline audio transcription logic with Home component for better code organization and maintainability
